v.0.0.8] feat. CharacterDeath 추가 중

+++ Assets/3.Script/Characters/Players/PlayerManagers/PlayerUIHudManager.cs
@@ -14,7 +14,7 @@ namespace KF
             healthBar.gameObject.SetActive(true);
         }
 
-        public void SetNewHeathValue(float oldValue, float newValue)
+        public void SetNewHeathValue(int oldValue, int newValue)
         {
             healthBar.SetStat(Mathf.RoundToInt(newValue));
         }
